fix: parser selection
