+++ src/grammar/parser-error.ts
@@ -1,7 +1,4 @@
-import { MatchResult } from "ohm-js";
 import { ErrorDisplay } from "../error/display";
-import { TactCompilationError } from "../errors";
-import { getSrcInfoFromOhm, ItemOrigin, SrcInfo } from "./src-info";
 
 const attributeSchema =
     (name: string) =>
@@ -112,23 +109,4 @@ export const syntaxErrorSchema = <T, U>(
 
 export type SyntaxErrors<T> = ReturnType<typeof syntaxErrorSchema<unknown, T>>;
 
-/**
- * @deprecated
- */
-export const parserErrorSchema = (display: ErrorDisplay<string>) => ({
-    ...syntaxErrorSchema(display, (message) => (source: SrcInfo) => {
-        throw new TactCompilationError(display.at(source, message), source);
-    }),
-    generic: (matchResult: MatchResult, path: string, origin: ItemOrigin) => {
-        const interval = matchResult.getInterval();
-        const source = getSrcInfoFromOhm(interval, path, origin);
-        // eslint-disable-next-line @typescript-eslint/no-explicit-any
-        const message = `Expected ${(matchResult as any).getExpectedText()}\n`;
-        throw new TactCompilationError(display.at(source, message), source);
-    },
-});
 
-/**
- * @deprecated
- */
-export type ParserErrors = ReturnType<typeof parserErrorSchema>;
