+++ src/grammar/prev/src-info.ts
@@ -1,0 +1,13 @@
+import { Interval } from "ohm-js";
+import { getSrcInfo, ItemOrigin, SrcInfo } from "../src-info";
+
+/**
+ * @deprecated
+ */
+export const getSrcInfoFromOhm = (
+    { sourceString, startIdx, endIdx }: Interval,
+    file: string | null,
+    origin: ItemOrigin,
+): SrcInfo => {
+    return getSrcInfo(sourceString, startIdx, endIdx, file, origin);
+};
