+++ src/grammar/test/expr-equality.spec.ts
@@ -1,5 +1,6 @@
 import { eqExpressions, getAstFactory } from "../ast";
 import { getParser } from "../";
+import { defaultParser } from "../grammar";
 
 type Test = { expr1: string; expr2: string; equality: boolean };
 
@@ -367,7 +368,7 @@ const initOfExpressions: Test[] = [
 
 function testEquality(expr1: string, expr2: string, equal: boolean) {
     const ast = getAstFactory();
-    const { parseExpression } = getParser(ast);
+    const { parseExpression } = getParser(ast, defaultParser);
     expect(eqExpressions(parseExpression(expr1), parseExpression(expr2))).toBe(
         equal,
     );
