+++ src/imports/resolveImports.spec.ts
@@ -3,6 +3,7 @@ import { createNodeFileSystem } from "../vfs/createNodeFileSystem";
 import path from "path";
 import { getParser } from "../grammar";
 import { getAstFactory } from "../grammar/ast";
+import { defaultParser } from "../grammar/grammar";
 
 describe("resolveImports", () => {
     it("should resolve imports", () => {
@@ -17,7 +18,7 @@ describe("resolveImports", () => {
             project,
             stdlib,
             entrypoint: "./main.tact",
-            parser: getParser(ast),
+            parser: getParser(ast, defaultParser),
         });
         expect(resolved).toMatchObject({
             func: [
